+++ lib/kafka/ffi/admin/delete_groups_result.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "kafka/ffi/event"
 require "kafka/ffi/admin/topic_result"
 
 module Kafka::FFI::Admin
@@ -8,6 +7,8 @@ module Kafka::FFI::Admin
   #
   # @see #groups
   class DeleteGroupsResult < ::Kafka::FFI::Event
+    event_type :delete_groups
+
     def self.new(event)
       ::Kafka::FFI.rd_kafka_event_DeleteGroups_result(event)
     end