+++ lib/kafka/ffi/admin/delete_topics_result.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
-require "kafka/ffi/event"
 require "kafka/ffi/admin/topic_result"
 
 module Kafka::FFI::Admin
   class DeleteTopicsResult < ::Kafka::FFI::Event
+    event_type :delete_topics
+
     def self.new(event)
       ::Kafka::FFI.rd_kafka_event_DeleteTopics_result(event)
     end