+++ lib/kafka/ffi/event.rb
@@ -4,14 +4,63 @@ require "kafka/ffi/opaque_pointer"
 
 module Kafka::FFI
   class Event < OpaquePointer
-    # LogMessage is attached to RD_KAFKA_EVENT_LOG events.
-    LogMessage = Struct.new(:facility, :message, :level) do
-      # @attr facility [String] Log facility
-      # @attr message [String] Log message
-      # @attr level [Integer] Verbosity level of the message
+    class << self
+      # registry holds mappings of librdkafka Event constants (RD_KAFKA_EVENT_*)
+      # to the class for that event. Classes are registered by calling the class
+      # method event_type.
+      def registry
+        @registry ||= {}
+      end
 
-      def to_s
-        message
+      # Registers the current event subclass to map to the given enum value.
+      #
+      # @see ffi.rb: enum :event_type
+      #
+      # @param type [Symbol] event type enum value
+      def event_type(type)
+        if Event.registry.key?(type)
+          raise ArgumentError, "#{type} already mapped to #{Event.registry[type]}"
+        end
+
+        Event.registry[type] = self
+      end
+
+      def from_native(ptr, _ctx)
+        if !ptr.is_a?(::FFI::Pointer)
+          raise TypeError, "from_native can only convert from a ::FFI::Pointer to #{self}"
+        end
+
+        # The equivalent of a native NULL pointer is nil.
+        if ptr.null?
+          return nil
+        end
+
+        type  = ::Kafka::FFI.rd_kafka_event_type(ptr)
+        klass = Event.registry[type] || Event
+
+        obj = klass.allocate
+        obj.send(:initialize, ptr)
+        obj
+      end
+
+      # @param value [Opaque, nil]
+      #
+      # @return [FFI::Pointer] Pointer referencing the Event
+      def to_native(value, _ctx)
+        if value.nil?
+          return ::FFI::Pointer::NULL
+        end
+
+        # Allow a ::FFI::Pointer to be passed through
+        if value.is_a?(::FFI::Pointer)
+          return value
+        end
+
+        if !value.is_a?(self)
+          raise TypeError, "expected a kind of #{self}, was #{value.class}"
+        end
+
+        value.pointer
       end
     end
 
@@ -29,47 +78,6 @@ module Kafka::FFI
     # @return [String] Name of the type of event
     def name
       ::Kafka::FFI.rd_kafka_event_name(self)
-    end
-
-    # Retrieve the set of messages attached to the event.
-    #
-    # Events:
-    #   - RD_KAFKA_EVENT_FETCH
-    #   - RD_KAFKA_EVENT_DR
-    #
-    # @note Do not call #destroy on the Messages
-    #
-    # @return [Array<Message>] Messages attached to the Event
-    def messages
-      count = ::Kafka::FFI.rd_kafka_event_message_count(self)
-      if count == 0
-        return []
-      end
-
-      begin
-        # Allocates enough memory for the full set but only converts as many
-        # as were returned.
-        # @todo Retrieve all until sum(ret) == count?
-        ptr = ::FFI::MemoryPointer.new(:pointer, count)
-        ret = ::Kafka::FFI.rd_kafka_event_message_array(self, ptr, count)
-
-        # Map the return pointers to Messages
-        return ptr.read_array_of_pointer(ret).map! { |p| Message.new(p) }
-      ensure
-        ptr.free
-      end
-    end
-
-    # Returns the configuration for the event or nil if the configuration
-    # property is not set or not applicable for the event type.
-    #
-    # Events:
-    #   - RD_KAFKA_EVENT_OAUTHBEARER_TOKEN_REFRESH
-    #
-    # @return [String] Configuration string for the event
-    # @return [nil] Property not set or not applicable.
-    def config_string
-      ::Kafka::FFI.rd_kafka_event_config_string(self)
     end
 
     # Returns the error code for the event or nil if there was no error.
@@ -110,89 +118,10 @@ module Kafka::FFI
 
     # Returns a description of the error or nil when there is no error.
     #
-    # @return [nil] No error for the Event
-    # @return [String] Description of the error
+    # @return [String, nil] Description of the error or nil if the event did
+    #   not have an error.
     def error_string
       ::Kafka::FFI.rd_kafka_event_error_string(self)
-    end
-
-    # Returns true or false if the Event represents a fatal error.
-    #
-    # @return [Boolean] There is an error for the Event and it is fatal.
-    def error_is_fatal
-      error && ::Kafka::FFI.rd_kafka_event_error_is_fatal(self)
-    end
-    alias error_is_fatal? error_is_fatal
-
-    # Returns the log message attached to the event.
-    #
-    # Events:
-    #   - RD_KAFKA_EVENT_LOG
-    #
-    # @return [Event::LogMessage] Attach log entry
-    def log
-      facility = ::FFI::MemoryPointer.new(:pointer)
-      message  = ::FFI::MemoryPointer.new(:pointer)
-      level    = ::FFI::MemoryPointer.new(:pointer)
-
-      exists = ::Kafka::FFI.rd_kafka_event_log(self, facility, message, level)
-      if exists != 0
-        # Event type does not support log messages.
-        return nil
-      end
-
-      LogMessage.new(
-        facility.read_pointer.read_string,
-        message.read_pointer.read_string,
-        level.read_int,
-      )
-    ensure
-      facility.free
-      message.free
-      level.free
-    end
-
-    # Extracts stats from the event
-    #
-    # Events:
-    #   - RD_KAFKA_EVENT_STATS
-    #
-    # @return [nil] Event type does not support stats
-    # @return [String] JSON encoded stats information.
-    def stats
-      # Calling stats on an unsupported type causes a segfault with librdkafka
-      # 1.3.0.
-      if type != :stats
-        return nil
-      end
-
-      ::Kafka::FFI.rd_kafka_event_stats(self)
-    end
-
-    # Returns the topic partition list from the Event.
-    #
-    # @note Application MUST NOT call #destroy on the list
-    #
-    # Events:
-    #   - RD_KAFKA_EVENT_REBALANCE
-    #   - RD_KAFKA_EVENT_OFFSET_COMMIT
-    #
-    # @return [TopicPartitionList, nil]
-    def topic_partition_list
-      tpl = ::Kafka::FFI.rd_kafka_event_topic_partition_list(self)
-      tpl.null? ? nil : tpl
-    end
-
-    # Returns the topic partition from the Event.
-    #
-    # @note The application MUST call #destroy on the TopicPartition when done.
-    #
-    # Events:
-    #   - RD_KAFKA_EVENT_ERROR
-    #
-    # @return [TopicPartition]
-    def topic_partition
-      ::Kafka::FFI.rd_kafka_event_topic_partition(self)
     end
 
     # Destroy the event, releasing it's resources back to the system.